fixing "bugs/errors" presented by openBSD

+++ Project1-3530/ssll.h
@@ -1,7 +1,8 @@
 #ifndef _SSLL_H_
 #define _SSLL_H_
-
-
+#include <cstddef>
+#include <exception>
+#include <new> 
 
 /*
 Skeleton created by Dave Small
@@ -217,7 +218,7 @@ namespace cop3530 {
 
 		//inserts elemenet, and shifts all elements after to the "left"
 		void insert(const T& element, int position) {
-			if (position < 0 || position > amount){
+			if (position < 0 || position > (int)amount){
 				if (position < 0){ throw std::domain_error("Domain Incorrect: does not take negative integers."); }
 				else{ throw std::domain_error("Domain Incorrect: Position does not exist."); }
 			}
